test1backend/requirements.txt
fastapi
uvicorn
google-cloud-vision
openai
supabase
python-dotenv